[libc][malloc] Make realloc grow and shrink in place
This increases performance, but doing so unconditionally would increase
fragmentation, since the best-fit policy would not be applied to the new
allocation. Accordingly, this only resizes in-place if the results agree
with that that would be produced by a free/malloc/memcpy. In particular,
for in-place reallocation to occur:

- The block must not have free space on both sides.
- The block that would be produced by freeing and coalescing the block
  must be the best fit block for the new size.

TODO: Actually implement the above.